ml_pipeline/pyproject.toml
[project]
name = "ml-pipeline"
version = "0.1.0"
description = "ML pipeline for ChocoForestWatch"
authors = [
    {name = "Luke Browne",email = "lukembrowne@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "scikit-learn",
    "pandas",
    "joblib",
    "python-dotenv",
    "SQLAlchemy",
    "GeoAlchemy2",
    "psycopg2-binary",
    "rasterio",
    "geopandas",
    "pyproj",
    "shapely",
    "tqdm",
    "ipykernel (>=6.29.5,<7.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "rio-tiler (>=7.7.0,<8.0.0)",
    "cogeo-mosaic (>=8.2.0,<9.0.0)",
    "boto3 (>=1.38.15,<2.0.0)",
    "pystac (>=1.13.0,<2.0.0)",
    "rio-stac (>=0.11.0,<0.12.0)",
    "pypgstac (>=0.9.6,<0.10.0)",
    "psycopg (>=3.2.9,<4.0.0)",
    "psycopg-pool (>=3.2.6,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "leafmap (>=0.45.0,<0.46.0)",
    "xarray (>=2025.4.0,<2026.0.0)",
    "localtileserver (>=0.10.6,<0.11.0)",
    "setuptools (>=80.4.0,<81.0.0)",
    "xgboost (>=3.0.1,<4.0.0)"
]

[project.optional-dependencies]
dev = [
    "jupyterlab",
    "ipykernel"
]

[tool.poetry]
name = "ml-pipeline"
version = "0.1.0"
description = "ML pipeline for ChocoForestWatch"
authors = ["Luke Browne <lukembrowne@gmail.com>"]
packages = [
    { include = "ml_pipeline", from = "src" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
